initial commit with a untrained detector
Need to train the detector on either a dataset to evaluate its performance or on the given dataset.

Data should hold corresponding paths to the dataset, and preprocessing_data.py should hold the steps to preprocess the data ( converting videos to image frames and getting the bounding boxes from the CSVs.